# docs/swagger.yaml
basePath: /
definitions:
  handler.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  handler.CreatePaymentRequest:
    properties:
      order_id:
        example: 123
        type: integer
      provider:
        example: momo
        type: string
    type: object
  handler.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  handler.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
      username:
        example: johndoe
        type: string
    type: object
  handler.ResetPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
      new_password:
        example: newpassword123
        type: string
      token:
        example: RESET123
        type: string
    type: object
...
  handler.UploadResponse:
    properties:
      url:
        type: string
    type: object
  model.CartItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      user_id:
        type: integer
    type: object
  model.Order:
    properties:
      address:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      customer_phone:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      payment_method:
        type: string
      shipping_method:
        type: string
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  model.OrderItem:
    properties:
      price:
        type: number
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
    type: object
  model.Product:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      status:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.Voucher:
    properties:
      code:
        description: Mã voucher, e.g., "SUMMER2026"
        type: string
      created_at:
        type: string
      description:
        description: Mô tả voucher
        type: string
      end_date:
        description: Ngày kết thúc
        type: string
      id:
        type: integer
      is_active:
        description: Còn hoạt động không
        type: boolean
      max_discount:
        description: Giảm tối đa (cho percentage)
        type: number
      min_order:
        description: Đơn hàng tối thiểu
        type: number
      start_date:
        description: Ngày bắt đầu
        type: string
      type:
        description: '"percentage" hoặc "fixed"'
        type: string
      usage_limit:
        description: Số lần sử dụng tối đa
        type: integer
      used_count:
        description: Đã dùng bao nhiêu lần
        type: integer
      value:
        description: Giá trị giảm (% hoặc số tiền)
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Clothing Store.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Clothing Store API
  version: "1.0"
paths:
  /cart:
    get:
      consumes:
      - application/json
      description: Get all items in the current user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CartItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: Add a product to the user's cart or increase quantity
      parameters:
      - description: Product and Quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.AddToCartRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: Added to cart
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - cart
...
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Request a password reset token via email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message, token (for demo)
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Request password reset
      tags:
      - auth
  /login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Login user
      tags:
      - auth
  /my/orders:
    get:
      consumes:
      - application/json
      description: Get a paginated list of orders for the currently logged-in user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List my orders
      tags:
      - orders
  /orders:
    post:
      consumes:
      - application/json
      description: Create a new order with items
      parameters:
      - description: Order Data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
...
  /payments/create:
    post:
      consumes:
      - application/json
      description: Create a payment URL for an order
      parameters:
      - description: Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create payment URL
      tags:
      - payments
  /payments/history:
    get:
      consumes:
      - application/json
      description: Get payment history
      produces:
      - application/json
      responses:
        "200":
          description: Not implemented yet
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get payment history
      tags:
      - payments
  /products:
    get:
      consumes:
      - application/json
      description: Get a list of products with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product (Admin only)
      parameters:
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - products
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register user
      tags:
      - auth
  /reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using the provided token
      parameters:
      - description: Reset details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            type: string
        "400":
          description: Invalid token or request
          schema:
            type: string
      summary: Reset password
      tags:
      - auth
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file and get the URL
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UploadResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload an image
      tags:
      - upload
  /vouchers:
    get:
      consumes:
      - application/json
      description: Get a list of vouchers with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Show only active vouchers
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Voucher'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List vouchers
      tags:
      - vouchers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"